feat(engine): make batch jobs deployment-aware
* generalize deployment-awareness in batch jobs creation
* remove coupling to job handlers from batch monitor handler

+++ engine/src/main/java/org/camunda/bpm/engine/impl/batch/BatchEntity.java
@@ -243,7 +243,7 @@ public class BatchEntity implements Batch, DbEntity, HasDbReferences, Nameable, 
 
   @Override
   public Object getPersistentState() {
-    HashMap<String, Object> persistentState = new HashMap<String, Object>();
+    HashMap<String, Object> persistentState = new HashMap<>();
     persistentState.put("jobsCreated", jobsCreated);
     return persistentState;
   }
@@ -333,12 +333,14 @@ public class BatchEntity implements Batch, DbEntity, HasDbReferences, Nameable, 
     }
   }
 
-  public void delete(boolean cascadeToHistory) {
+  public void delete(boolean cascadeToHistory, boolean deleteJobs) {
     CommandContext commandContext = Context.getCommandContext();
 
     deleteSeedJob();
     deleteMonitorJob();
-    getBatchJobHandler().deleteJobs(this);
+    if (deleteJobs) {
+      getBatchJobHandler().deleteJobs(this);
+    }
 
     JobDefinitionManager jobDefinitionManager = commandContext.getJobDefinitionManager();
     jobDefinitionManager.delete(getSeedJobDefinition());
@@ -403,13 +405,13 @@ public class BatchEntity implements Batch, DbEntity, HasDbReferences, Nameable, 
 
   @Override
   public Set<String> getReferencedEntityIds() {
-    Set<String> referencedEntityIds = new HashSet<String>();
+    Set<String> referencedEntityIds = new HashSet<>();
     return referencedEntityIds;
   }
 
   @Override
   public Map<String, Class> getReferencedEntitiesIdAndClass() {
-    Map<String, Class> referenceIdAndClass = new HashMap<String, Class>();
+    Map<String, Class> referenceIdAndClass = new HashMap<>();
 
     if (seedJobDefinitionId != null) {
       referenceIdAndClass.put(seedJobDefinitionId, JobDefinitionEntity.class);
